refactor: move redis api impl into standalone file
